requirements.txt
opencv-python~=4.4.0.46
requests~=2.25.1
numpy~=1.19.4
matplotlib~=3.3.3
mtcnn~=0.1.0
Pillow~=8.0.1
torchvision~=0.8.2
torch~=1.7.1